refactor: Split oversized files main.c and orchestrator.c (Issue #11)
main.c refactoring (1791 → 503 lines, 72% reduction):
- Extract src/core/commands/commands.c (1061 lines) - All cmd_* handlers
- Extract src/core/repl.c (554 lines) - REPL loop, input processing
- Extract src/core/signals.c (89 lines) - Signal handling

orchestrator.c refactoring (1620 → 1300 lines, 20% reduction):
- Extract src/orchestrator/delegation.c (263 lines) - Agent task routing
- Extract src/orchestrator/planning.c (67 lines) - Execution plans
- Extract src/orchestrator/convergence.c (67 lines) - Response synthesis

New headers:
- include/nous/commands.h - Command handler declarations
- include/nous/repl.h - REPL function declarations
- include/nous/signals.h - Signal handling declarations
- include/nous/delegation.h - Delegation API
- include/nous/planning.h - Planning API
- include/nous/convergence.h - Convergence API

Benefits:
- Improved modularity and maintainability
- Better separation of concerns
- Easier testing of individual components
- Follows repository coding standards

Closes #11
