+++ src/orchestrator/cost.c
@@ -175,7 +175,8 @@ double cost_get_remaining_budget(void) {
     if (!orch || orch->cost.budget_limit_usd <= 0) return -1.0;  // No budget set
 
     CONVERGIO_MUTEX_LOCK(&g_cost_mutex);
-    double remaining = orch->cost.budget_limit_usd - orch->cost.current_spend_usd;
+    // Use total spend, not just session spend
+    double remaining = orch->cost.budget_limit_usd - orch->cost.total_spend_usd;
     CONVERGIO_MUTEX_UNLOCK(&g_cost_mutex);
 
     return remaining > 0 ? remaining : 0.0;
@@ -191,7 +192,8 @@ void cost_set_budget(double limit_usd) {
 
     CONVERGIO_MUTEX_LOCK(&g_cost_mutex);
     orch->cost.budget_limit_usd = limit_usd;
-    orch->cost.budget_exceeded = (orch->cost.current_spend_usd >= limit_usd);
+    // Check budget against total spend, not just session
+    orch->cost.budget_exceeded = (orch->cost.total_spend_usd >= limit_usd);
     CONVERGIO_MUTEX_UNLOCK(&g_cost_mutex);
 }
 
@@ -229,49 +231,79 @@ char* cost_get_report(void) {
     time_t now = time(NULL);
     int session_minutes = (int)((now - orch->cost.session_start) / 60);
 
-    // Format budget status
-    char budget_status[128];
+    // Format budget status - use TOTAL spend, not just session
+    char budget_line[64];
     if (orch->cost.budget_limit_usd > 0) {
-        double pct = (orch->cost.current_spend_usd / orch->cost.budget_limit_usd) * 100;
-        snprintf(budget_status, sizeof(budget_status),
-            "Budget: $%.2f / $%.2f (%.1f%%)%s",
-            orch->cost.current_spend_usd,
-            orch->cost.budget_limit_usd,
-            pct,
-            orch->cost.budget_exceeded ? " [EXCEEDED]" : "");
+        double pct = (orch->cost.total_spend_usd / orch->cost.budget_limit_usd) * 100;
+        if (orch->cost.budget_exceeded) {
+            snprintf(budget_line, sizeof(budget_line),
+                "$%.2f / $%.2f (%.0f%%) EXCEEDED",
+                orch->cost.total_spend_usd,
+                orch->cost.budget_limit_usd,
+                pct);
+        } else {
+            snprintf(budget_line, sizeof(budget_line),
+                "$%.2f / $%.2f (%.0f%%)",
+                orch->cost.total_spend_usd,
+                orch->cost.budget_limit_usd,
+                pct);
+        }
     } else {
-        snprintf(budget_status, sizeof(budget_status), "Budget: No limit set");
+        snprintf(budget_line, sizeof(budget_line), "No limit set");
     }
 
-    snprintf(report, 2048,
-        "╔══════════════════════════════════════════════════════════════╗\n"
-        "║                    COST REPORT                               ║\n"
-        "╠══════════════════════════════════════════════════════════════╣\n"
-        "║ SESSION (%d min)                                              \n"
-        "║   Input tokens:  %12llu ($%.4f)                             \n"
-        "║   Output tokens: %12llu ($%.4f)                             \n"
-        "║   Total cost:    $%.4f                                       \n"
-        "╠══════════════════════════════════════════════════════════════╣\n"
-        "║ ALL-TIME                                                     \n"
-        "║   Input tokens:  %12llu ($%.4f)                             \n"
-        "║   Output tokens: %12llu ($%.4f)                             \n"
-        "║   Total cost:    $%.4f                                       \n"
-        "╠══════════════════════════════════════════════════════════════╣\n"
-        "║ %s\n"
-        "╚══════════════════════════════════════════════════════════════╝\n",
+    // Build report with properly aligned borders
+    int offset = 0;
+    offset += snprintf(report + offset, 2048 - offset,
+        "\n\033[1m╔════════════════════════════════════════════════════╗\033[0m\n"
+        "\033[1m║               COST REPORT                          ║\033[0m\n"
+        "\033[1m╠════════════════════════════════════════════════════╣\033[0m\n");
+
+    // Session section
+    offset += snprintf(report + offset, 2048 - offset,
+        "\033[36m║ SESSION\033[0m (%d min)\n"
+        "║   Input tokens:  %'12llu  ($%.4f)\n"
+        "║   Output tokens: %'12llu  ($%.4f)\n"
+        "║   \033[1mTotal cost:      $%.4f\033[0m\n",
         session_minutes,
         (unsigned long long)orch->cost.session_usage.input_tokens,
         (orch->cost.session_usage.input_tokens / 1000000.0) * CLAUDE_SONNET_INPUT_COST,
         (unsigned long long)orch->cost.session_usage.output_tokens,
         (orch->cost.session_usage.output_tokens / 1000000.0) * CLAUDE_SONNET_OUTPUT_COST,
-        orch->cost.session_usage.estimated_cost,
+        orch->cost.session_usage.estimated_cost);
+
+    offset += snprintf(report + offset, 2048 - offset,
+        "\033[1m╠════════════════════════════════════════════════════╣\033[0m\n");
+
+    // All-time section
+    offset += snprintf(report + offset, 2048 - offset,
+        "\033[36m║ ALL-TIME\033[0m\n"
+        "║   Input tokens:  %'12llu  ($%.4f)\n"
+        "║   Output tokens: %'12llu  ($%.4f)\n"
+        "║   \033[1mTotal cost:      $%.4f\033[0m\n",
         (unsigned long long)orch->cost.total_usage.input_tokens,
         (orch->cost.total_usage.input_tokens / 1000000.0) * CLAUDE_SONNET_INPUT_COST,
         (unsigned long long)orch->cost.total_usage.output_tokens,
         (orch->cost.total_usage.output_tokens / 1000000.0) * CLAUDE_SONNET_OUTPUT_COST,
-        orch->cost.total_usage.estimated_cost,
-        budget_status
-    );
+        orch->cost.total_usage.estimated_cost);
+
+    offset += snprintf(report + offset, 2048 - offset,
+        "\033[1m╠════════════════════════════════════════════════════╣\033[0m\n");
+
+    // Budget section
+    if (orch->cost.budget_exceeded) {
+        offset += snprintf(report + offset, 2048 - offset,
+            "\033[31m║ BUDGET: %s\033[0m\n", budget_line);
+    } else if (orch->cost.budget_limit_usd > 0) {
+        offset += snprintf(report + offset, 2048 - offset,
+            "\033[32m║ BUDGET: %s\033[0m\n", budget_line);
+    } else {
+        offset += snprintf(report + offset, 2048 - offset,
+            "║ BUDGET: %s\n", budget_line);
+    }
+
+    offset += snprintf(report + offset, 2048 - offset,
+        "\033[1m╚════════════════════════════════════════════════════╝\033[0m\n");
 
     CONVERGIO_MUTEX_UNLOCK(&g_cost_mutex);
